requirements.txt
pandas
Pillow
requests
schedule
moviepy==2.1.1
google-genai
python-dotenv
numpy
